fixed:通过jwt对请求用户认证

+++ backend_todolist/src/main/java/com/todolistbackend/controller/TodoController.java
@@ -4,12 +4,14 @@ import com.alibaba.fastjson.JSON;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.todolistbackend.entity.Todo;
 import com.todolistbackend.mapper.TodoMapper;
+import com.todolistbackend.utils.TokenUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.List;
+import java.util.Map;
 
 @RestController
 public class TodoController {
@@ -17,30 +19,41 @@ public class TodoController {
     private TodoMapper todoMapper;
 
     @GetMapping("/todos/{userid}")
-    public String getTodos(@PathVariable Integer userid) {
-        QueryWrapper<Todo> wrapper = new QueryWrapper<>();
-        wrapper.eq("userid", userid);
-        List<Todo> todos = todoMapper.selectList(wrapper);
+    public String getTodos(@PathVariable Integer userid, @RequestHeader Map<String, String> head) {
+        List<Todo> todos = null;
+        if (TokenUtils.checkReq(head.get("token"), userid)) {
+            QueryWrapper<Todo> wrapper = new QueryWrapper<>();
+            wrapper.eq("userid", userid);
+            todos = todoMapper.selectList(wrapper);
+        }
         return JSON.toJSONString(todos);
     }
 
     @PostMapping("/delete")
-    public void delTodo(@RequestBody Todo todo) {
-        todoMapper.deleteById(todo);
+    public void delTodo(@RequestBody Todo todo, @RequestHeader Map<String, String> head) {
+        if (TokenUtils.checkReq(head.get("token"), todo.getUserid())) {
+            todoMapper.deleteById(todo);
+        }
     }
 
     @GetMapping("/count_total/{userid}")
-    public long countTodos(@PathVariable Integer userid) {
-        QueryWrapper<Todo> wrapper = new QueryWrapper<>();
-        wrapper.eq("userid", userid);
-        return culTodayTodo(wrapper);
+    public long countTodos(@PathVariable Integer userid, @RequestHeader Map<String, String> head) {
+        if (TokenUtils.checkReq(head.get("token"), userid)) {
+            QueryWrapper<Todo> wrapper = new QueryWrapper<>();
+            wrapper.eq("userid", userid);
+            return culTodayTodo(wrapper);
+        }
+        return 0;
     }
 
     @GetMapping("/count_finish/{userid}")
-    public long countFinishTodos(@PathVariable Integer userid) {
-        QueryWrapper<Todo> wrapper = new QueryWrapper<>();
-        wrapper.eq("status", 1).eq("userid", userid);
-        return culTodayTodo(wrapper);
+    public long countFinishTodos(@PathVariable Integer userid, @RequestHeader Map<String, String> head) {
+        if (TokenUtils.checkReq(head.get("token"), userid)) {
+            QueryWrapper<Todo> wrapper = new QueryWrapper<>();
+            wrapper.eq("status", 1).eq("userid", userid);
+            return culTodayTodo(wrapper);
+        }
+        return 0;
     }
 
     private long culTodayTodo(QueryWrapper<Todo> wrapper) {
@@ -57,13 +70,16 @@ public class TodoController {
     }
 
     @PostMapping("/add")
-    public void addTodo(@RequestBody Todo todo) {
-        todoMapper.insert(todo);
+    public void addTodo(@RequestBody Todo todo, @RequestHeader Map<String, String> head) {
+        if (TokenUtils.checkReq(head.get("token"), todo.getUserid())) {
+            todoMapper.insert(todo);
+        }
     }
 
     @PostMapping("/update")
-    public String updateTodo(@RequestBody Todo todo) {
-        todoMapper.updateById(todo);
-        return JSON.toJSONString(todo);
+    public void updateTodo(@RequestBody Todo todo, @RequestHeader Map<String, String> head) {
+        if (TokenUtils.checkReq(head.get("token"), todo.getUserid())) {
+            todoMapper.updateById(todo);
+        }
     }
 }
